refactor(use-cases): organize use cases by entity
