+++ src/domain/usecases/products/create-product-usecase.js
@@ -1,4 +1,4 @@
-const { MissingParamError } = require('./../../utils/errors');
+const { MissingParamError } = require('../../../utils/errors');
 
 class CreateProductUsecase {
   constructor({ createProductRepository, createProductFactory } = {}) {
