+++ src/main/composers/products/list-products-router-composer.js
@@ -1,5 +1,5 @@
 const ListProductsRouter = require('../../../presentation/routers/products/list-products-router');
-const ListProductsUseCase = require('../../../domain/usecases/list-products-usecase');
+const ListProductsUseCase = require('../../../domain/usecases/products/list-products-usecase');
 const ListProductsRepository = require('../../../infra/repositories/products/list-products-repository');
 const RequestValidator = require('../../../presentation/validators/request-validator');
 const { listProductsValidatorSchema } = require('../../../presentation/validators/schemas');
